Changes and Ui done

+++ composeApp/src/commonMain/kotlin/com/raunak/weatherforecast/geoLocation/data/mapper/GeoLocationMapper.kt
@@ -17,7 +17,7 @@ class GeoLocationMapper : Mapper<GeoLocation, GeoLocationEntity> {
                 countryCode,
                 countryId,
                 timezone,
-                elevation,
+                elevation.toDouble(),
                 flagUrl = K.flagUrl(countryCode)
             )
         }
